fixed mistake with the tensor type declaration

+++ aquila_inference.cc
@@ -270,7 +270,7 @@ void AquilaServiceImpl::DoRegressInBatch(
   }
 
   // Transform protobuf input to inference input tensor.
-  tensorflow::Tensor input(tensorflow::DT_FLOAT, {batch_size, kImageDataSize});
+  tensorflow::Tensor input(tensorflow::DT_UINT8, {batch_size, kImageDataSize});
   auto dst = input.flat_outer_dims<float>().data();
   // Assemble the batch into a tensor, copying it from the batch data to
   // the input tensor at location dst repeatedly for each item in the batch
